remove force refresh
Removed the force refresh, even though it looks kind of odd with a reload. When sitting on the home page already, and logging out, there's just a split second of button change in the navbar, so it's not a very visually pleasing log out. The reload made it clear that we logged out. Easy to put back in if we wish though